DEV-1262: Eclipse plugin should be consistent with module names. TC Eclipse plugin will no longer attempt to normalize the derived module name for the selected TIM jar file; eg: if the jar file is named baz-tim-1.0.0-SNAPSHOT.jar then it will now generate the ff config entry: <module name="baz-tim" version="1.0.0-SNAPSHOT" group-id="..."/> instead of: <module name="baz_tim" version="1.0.0.SNAPSHOT" group-id="..."/>
DEV-1263: Module loader can be improved in speed.
Module resolver is simplified. It now expects that the TIM jar file will be found either in the Maven-like path or the top of the repository and that the jar file name matches
the TIM symbolic name; eg: if version is 1.0.0 and symbolic name is org.foo.bar.tim_bazquux, then jar file is expected to be tim-bazquux-1.0.0.jar

DEV-1274: Eclipse plugin config editor needs to look inside the module to determine the actual module group-id, name, and version.
We don't need to do this anymore because of DEV-1262 and DEV-1263 fixes; read the comments in the JIRA entry for more info.



git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@6974 7fc7bbf3-cf45-46d4-be06-341739edd864