CDV-636: TC Version Check of TIM CDV-167: Add ability to check for dso version in config modules. DEV-1128: TIMs need manifest attribute specifying supported Terracotta version
Code and tests to check for required TC version for TIM's.
A new attribute named Terracotta-RequireVersion is added into the manifest of all TIM's that comes packaged with the kit. When TC installs a TIM at runtime, this attribute's vvalue is compared against TC's version number and appropriate action is taken (see CDV-636 for details)



git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@7467 7fc7bbf3-cf45-46d4-be06-341739edd864