Fixes an intermittent failure in BarrierIntegrationTest.
The recent change to no longer synchronize the entire await() method, in TerracottaBarrier, meant that it was now possible for multiple threads, from this single client, to run the server-side entity code, at the same time.
The barrier expects that all messages are delivered on one thread so this was causing lots of problems.
This is a simple change to force only one thread per entity to invoke into the entity.
While this is somewhat dirty, as it interacts with the monitor state of a foreign object, it is the closest to what we "mean", whereas the cleaner alternatives don't map to a single-threaded server, in any way (invocation builders are transient, end-points are per-client, etc) and simpler than introducing a new server-level concept into a test which will likely be migrated to our in-process system tests, later on.