TAB-6374 : Make client full stripe aware
Internally, pass our list of stripe members between components as a list of host:port pairs (potentially also including user names) instead of a comma-delimited list.
Note that many areas still require a flat string representation (for output, etc) so there are a few places where we re-serialize the list as a single string.
In many cases, this information is further encapsulated in TerracottaClientStripeConnectionConfig (prior to system property expansion).