Issue #350 Add more efficient shortest queue selection methods.
For StageQueueImpl, and interface StageQueue was extracted; a factory
is now used to create StageQueue's. This allowed me to split
StageQueueImpl into 2 different implementations. SingletonStageQueueImpl
is used when only 1 queue is desired (allowing for much more efficiency
in this case, while MultiStageQueueImpl is used for queue counts >1.

In MultiStageQueueImpl, I created an enum for two different shortest path
impls. BRUTE is the classic brute force approach, while PARTITION
simply partitions the queues into sets of 4 (or 2 if the number of
queue is <8, or 1 if <2), and then on each findShortestQueue(),
we choose the next partition, and brute force the shortest queue
within that partition. Since the partitions are so small, the short
brute force search is fast, and the clock-cache like behavior
of round robin selection of partition should be pretty good at finding
the shortest queue most of the time. Synthetic testing showed this to
be extremely efficient.

Another minheap approach was tried, but the overhead on put()/poll()
was too great to be useful.

A system property, "tc.stagequeueimpl.findstrategy", is available which
cane be set to one of BRUTE or PARTITION to choose the findstrategy
while we test in real world conditions. By default, PARTITION is used.