ClientSourceId for message tracker, notifyDestroy() support
Here we have added ClientSourceId to InvokeContext, and moved
ClientDescriptor to ActiveInvokeContext. Properly, message tracking
shoudl be done in terms of the ClientSourceId/txid/eldest tuple,
not ClientDescriptor.

ClientDescriptor is no longer serializable, as that is a mistake. It
is not a useful thing to have on a passive; when a passive fails over,
it will get a real ClientDescriptor via handleReconnect()

Added a notifyDestroyed(ClientSourceId) call on CommonServerEntity,
and therefore ManagedEntitImpl, with appropriate support. Currently,
all entities will be notified for all disconnects; safely ignore ones
which are not relevant. This can be used to spark cleanup of
no longer useful message tracking information.