Avoid IllegalStateException from MockWeakReference.get in ServerStackProviderTest
This commit adjusts creation of a mock in
ServerStackProviderTest.testAttachAlreadyConnectedTransport. There
appears to have been a race condition, possibly exacerbated by the
changes made for PortManager adoption, between creation of the mock
for NetworkStackHarnessFactory and obtaining a reference to that mock
using the 'getMock' method.  This code is changed to retain a direct
reference to the mock rather than relying on the 'getMock' method.