Updates to adjust for PortManager port release check
This commit updates the version of terracotta-utilities to adapt to the
PortManager port release checker and PortRef.close options.

This commit changes the 'galvan' module's test execution to suppress
PortManager release checking.  It also re-enables one skipped test that
had an increased false failure rate due to the port release checking.

This test also changes internal calls to PortRef.close to use the "option"
version to avoid the port release checker overhead in
AbstractHarnessEntry.chooseRandomPortRange.  (Port release checking
continues to be enabled when the assigned ports are released.)